AppForDeployment/requirements.txt
openai
pandas
numpy
python-dotenv
fastapi
requests
uvicorn[standard]
